npc commit
