+++ Assets/Imported/BodyGuards/Meshes/Materials.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: da052d6c6bc53dc448c83cb4ef2d3f41
+guid: 7747a8b2aa63a9e439b12f01f5ac8aac
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
