+++ Assets/Kevin Iglesias/Basic Motions.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2cb279c410b7a3e4dbf4aab42283effa
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
