+++ Assets/Kevin Iglesias/Basic Motions/Animations.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7eb11f9df068c2b45bc52cfc0d6e3f25
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
