+++ Assets/Kevin Iglesias/Basic Motions/Animations/Movement.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 162d3c986ae81044b8e9de5caae05de1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
