+++ Assets/Kevin Iglesias/Basic Motions/Models.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c310a300a0358f94f81e0eeb585d6cf4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
