+++ Assets/Prefab/WayPoint.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a5eefa85abbabdb48a2f2ff504369d8f
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
