+++ Assets/Prefab/npcPrefab.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 25692a6866db5144381e6f6bdaf689ec
+guid: 0fcc4ceaa5f947e47b669fd6448eeb0c
 PrefabImporter:
   externalObjects: {}
   userData:
