+++ Assets/QuarterView 3D Action BE5.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1e3914da9a64443468870d2e67277412
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
