+++ Assets/QuarterView 3D Action BE5/Models.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d7209273a7480ad4f95a1fa251e44664
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
