+++ Assets/Script/Npc.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c4126ca594b1ca34d9b205bd7aac5cad
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
