+++ Assets/Script/Npc/Enemy A.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3dc4fd980657e194bb8b71fa70d4736e
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
