+++ Assets/Script/Npc/NpcController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2a6bdd387abb0eb4881e3c6e47feb387
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
