+++ Assets/Script/Npc/NpcTest.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1f835179eb0d38f43b36a6f7b60d616a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
